refactor(core): Move DefaultVersionCalculator from CLI to core crate
DefaultVersionCalculator was only accessible from the CLI crate, which blocked
the server binary from using it without an import cycle. Moving it to core means
both CLI and server can build a ReleaseRegentProcessor with a concrete version
calculator without duplicating or cross-crate borrowing.

CLI version_calculator module now re-exports release_regent_core::DefaultVersionCalculator.
